pushed shared fields into own base class. cleaned up spotbugs file.
