+++ src/main/java/org/hyperledger/aries/api/exception/ParsingException.java
@@ -1,0 +1,14 @@
+/*
+ * Copyright (c) 2020-2022 - for information on the respective copyright owner
+ * see the NOTICE file and/or the repository at
+ * https://github.com/hyperledger-labs/acapy-java-client
+ *
+ * SPDX-License-Identifier: Apache-2.0
+ */
+package org.hyperledger.aries.api.exception;
+
+public class ParsingException extends RuntimeException {
+    public ParsingException(String message) {
+        super(message);
+    }
+}
